rag-service/requirements.txt
raganything
lightrag-hku
fastapi
uvicorn[standard]
python-multipart
google-genai
numpy
